# Optimize AR::Timestamp#clear_timestamp_attributes