@casperisfine 2024-10-31
Same benchmark, but with only the `.dup` and no `assign_attributes`

Before:
```
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
        ActiveRecord    14.473k i/100ms
Calculating -------------------------------------
        ActiveRecord    150.644k (± 3.8%) i/s    (6.64 μs/i) -    752.596k in   5.004512s
```

After:
```
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
        ActiveRecord    16.750k i/100ms
Calculating -------------------------------------
        ActiveRecord    170.844k (± 5.9%) i/s    (5.85 μs/i) -    854.250k in   5.022552s
```